+++ examples/type-aware-usage.ts
@@ -8,7 +8,7 @@
  */
 
 // eslint.config.ts
-import eslintRules from '@zayat/eslint-custom-rules';
+import eslintRules from 'zayat-eslint-rules';
 
 export default [
 	// Use the recommended configuration